Path A: narrow minic near pointers from 'l' to 'w' on i8086
Stevie's .COM dropped from 97982 → 80922 bytes (~17%) after switching
near pointer IR class from 32-bit (l) to 16-bit (w) for i8086 small
model. Far pointers stay l-typed.

minic.y: irtyp/irtyp_ret return 'w' for non-far PTR/FUN. Hard-coded
=l alloc/copy/add/mul/div sites that produce or consume near-pointer
addresses now emit =w. Function-pointer locals shrink from alloc8 8 to
alloc4 2. ptrdiff_t for near pointers is INT (16-bit) not LNG.

ops.h: alloc4/8/16 accept Kw results in addition to Kl, since on i8086
the alloc'd address is 16-bit.

parse.c: usecheck allows Kw tmp where Km (=Kl) is expected, gated on
the i8086 target so amd64/arm64 typechecking is unchanged.

i8086/emit.c: Kw constant operands are sign-truncated to 16 bits before
emission, so a folded Kw `sub 0,1` (= 0xFFFFFFFF in QBE's 32-bit-w
arithmetic) prints as -1 / 0xFFFF rather than blowing past NASM's word
bound.

stddef.h: size_t and ptrdiff_t become unsigned int / int (16-bit on
small model). Was unsigned long, which forced 32-bit-pair codegen for
every strlen()+1 expression.

libstub.asm: pointer-arg stack offsets shrink from 4-byte slots back to
2-byte (intdos outregs [bp+6], int86 outregs [bp+8], strrchr c [bp+6],
strcpy src [bp+6]).  _stdin/_stdout/_stderr globals added so fprintf()
references resolve.

24/24 stevie sources still compile and link; binary still over the
.COM 64KB ceiling, so a follow-up (Path B .EXE, or further codegen
shrinkage) is required to actually run end-to-end.

+++ minic/include/stddef.h
@@ -3,8 +3,11 @@
 #ifndef _STDDEF_H
 #define _STDDEF_H
 
-typedef unsigned long size_t;
-typedef long ptrdiff_t;
+/* On i8086 small/medium model, size_t and ptrdiff_t are 16-bit
+ * (unsigned int / int).  Using long here would force 32-bit pair ops
+ * for every string-length / sizeof-driven expression. */
+typedef unsigned int size_t;
+typedef int ptrdiff_t;
 
 /* MiniC's grammar doesn't yet accept `((void *)0)` at file scope as a
  * pointer initializer; both forms are valid null pointer constants in
